Amount changed from int64 to float64

+++ internal/service/transactions.go
@@ -8,7 +8,7 @@ import (
 )
 
 type TransactionsService interface {
-	AddPayment(userID int, msgID int, amount int64, description string, source string, timestamp int64) error
+	AddPayment(userID int, msgID int, amount float64, description string, source string, timestamp int64) error
 }
 
 type transactionsService struct {
@@ -21,7 +21,7 @@ func NewTransactionsService(repo repository.TransactionsRepository) Transactions
 	}
 }
 
-func (srv *transactionsService) AddPayment(userID int, msgID int, amount int64, description string, source string, timestamp int64) error {
+func (srv *transactionsService) AddPayment(userID int, msgID int, amount float64, description string, source string, timestamp int64) error {
 	// TODO Usar el msgID para actualizar la info en la DB en caso de que se edite el mensaje
 	transactionDTO := domain.TransactionDTO{
 		MsgID:       msgID,
